pyproject.toml
[project]
name = "moka"
version = "0.1.0"
description = "MokA: Multimodal Low-Rank Adaptation for MLLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.53.2",
    "deepspeed>=0.12.6",
    "accelerate",
    "peft",
    "datasets",
    "numpy",
    "pillow",
    "opencv-python",
    "librosa",
    "sentencepiece",
    "huggingface-hub",
    "safetensors",
    "tqdm",
    "pyyaml",
    "pip>=26.0.1",
    "yt-dlp>=2026.3.17",
]

[[index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
